style: :lipstick: Refactors and enhances project structure
This commit restructures and enhances the project for improved maintainability, scalability, and performance.  - Migrates to a monorepo structure with dedicated packages for the main app and design system to increase code reuse and separation of concerns - Updates build and configuration files to be compatibile with the new structure - Introduces Preact to the design system for faster rendering and smaller bundle size. - Implements new UI components for blog cards, footers, links, navigation menus and project cards - Adds configuration and tooling for linting, formatting, SEO, and static site generation